Icon loader error on startup https://bugs.webkit.org/show_bug.cgi?id=174787
Reviewed by Brady Eidson.

Source/WebCore:

This is a regression of the new icon loading, it happens with pages that shouldn't have a favicon, like about
pages. IconController::startLoader() did several checks before starting the load that
DocumentLoader::startIconLoading() is not doing. It checked that the frame is the main one, the document can have
an icon (document url is not empty and not about:blank) and that favicon url is in HTTP family. We should do the
same checks now before starting to load icons.

* loader/DocumentLoader.cpp:
(WebCore::DocumentLoader::startIconLoading):

Source/WebKit:

Ignore non HTTP favicons in glib API.

* UIProcess/API/glib/WebKitIconLoadingClient.cpp:

Identifier: 191665@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@219905 268f45cc-cd09-0410-ab3c-d52691b4dbfc