[GCrypt] Key serialization support https://bugs.webkit.org/show_bug.cgi?id=173883
Reviewed by Jiewen Tan.

Source/WebCore:

Implement crypto key serialization for platforms that use libgcrypt.

Despite the function names and the existing implementation for the Mac port,
we don't perform the KEK wrapping and unwrapping due to the review feedback
advising against it.

Instead the passed-in key data is simply copied and returned as the result of
the 'wrapping' operation, and similarly the passed-in serialized data is simply
copied and returned as the result of the 'unwrapping' operation.

No new tests -- related tests are now passing and are unskipped.

* crypto/gcrypt/SerializedCryptoKeyWrapGCrypt.cpp:
(WebCore::getDefaultWebCryptoMasterKey):
(WebCore::wrapSerializedCryptoKey):
(WebCore::unwrapSerializedCryptoKey):

LayoutTests:

* platform/wpe/TestExpectations: Unskip the crypto/worker/ tests that
utilize key serialization and are now passing.


Identifier: 191705@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@219976 268f45cc-cd09-0410-ab3c-d52691b4dbfc