Remove unused and obsolete setting mediaDocumentEntersFullscreenAutomatically https://bugs.webkit.org/show_bug.cgi?id=175080
Patch by Jeremy Jones <jeremyj@apple.com> on 2017-08-02
Reviewed by Jon Lee.

Source/WebCore:

No new tests because this only removes unused code.

This is obsolete because of the alternate solution in
https://bugs.webkit.org/show_bug.cgi?id=174850

* page/Settings.in:

Source/WebKit:

This is obsolete because of the alternate solution in
https://bugs.webkit.org/show_bug.cgi?id=174850

* Shared/WebPreferencesDefinitions.h:
* UIProcess/API/Cocoa/WKPreferences.mm:
(-[WKPreferences _setMediaDocumentEntersFullscreenAutomatically:]): Deleted.
(-[WKPreferences _mediaDocumentEntersFullscreenAutomatically]): Deleted.
* UIProcess/API/Cocoa/WKPreferencesPrivate.h:
* WebProcess/WebPage/WebPage.cpp:
(WebKit::WebPage::updatePreferences):

Identifier: 191852@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@220151 268f45cc-cd09-0410-ab3c-d52691b4dbfc