WKPreferences should conform to NSCopying https://bugs.webkit.org/show_bug.cgi?id=175303 <rdar://problem/33764939>
Reviewed by Dan Bernstein.

Source/WebKit:

* UIProcess/API/Cocoa/WKPreferencesPrivate.h:
Add conformance in the private header, as there is not consensus that this
needs to be API.

* UIProcess/API/Cocoa/WKPreferences.mm:
(-[WKPreferences copyWithZone:]):
Copy the wrapped WebPreferences instance and return its wrapper.

Tools:

Add unit test to make sure a WKPreference and its copy are
not aliased to each other or affect each other's properties.

* TestWebKitAPI/TestWebKitAPI.xcodeproj/project.pbxproj:
* TestWebKitAPI/Tests/WebKit2Cocoa/Copying.mm: Added.
(TEST):


Identifier: 192150@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@220542 268f45cc-cd09-0410-ab3c-d52691b4dbfc